[Chore] Battle 내 유물 연결
* Battle 내 유물 연결
